rdma: Add functions to support memory translation
This is the first commit in a series which aim is to
unify MRs registration and memory translation in
NVMEoF target and NVME initiator RDMA transports.

This commit also adds a function to register
some custom memory translation callback. This
callback is different from approach implemented with
nvme_rdma hooks since it allows to perfrom memory
a custom translation in runtime.

Next patches will add usage of new functions in
RDMA transports.

Change-Id: I93153d497fc4554ced14edbe545961b78bda91e3
Signed-off-by: Alexey Marchuk <alexeymar@mellanox.com>