bdev: Add new IO type DATA_PASSTHRU
If bdev module supports this IO type that means
the the module doesn't read or modify data passed
in IO request.
For now only bdev_nvme which uses RDMA transport
supports the new IO type.

In some cases IO data buffers may not belong to a host
memory (e.g. GPUDirect). It can be useful for the
application to understand if bdev needs data to be
located in local CPU memory, so that the application
should copy data or it is possible to skip unnecessary
data copy.

Change-Id: Ia40222ab8c521442515a4aca5ff86bd38376c4fb
Signed-off-by: Alexey Marchuk <alexeymar@mellanox.com>