rdma: Add memory translation callback
This commit adds a function to register custom memory
translation callback. This callback differs from approach
implemented with nvme_rdma hooks since it allows to perfrom
memory translation in runtime.

In some scenarios it might be needed to create a new
Memory Region that won't be included in the standard
SPDK memory translation mechanism but RDMA transport
implementation still need to access it.

Change-Id: I2bff99c6e0a97c014df2ab5165a2bf7bd7f0d5d1