bdev/rbd: Loadshare rbd bdevs between CPU cores
In current implementation all IOs for bdev rbd are served by single
"main" thread. It is set to the thread that created the first IO
channel for this rbd bdev. In many SPDK apps it will be the same
thread for all rbd bdevs, e.g. the 'nvmf_tgt_poll_group_0' for
nvmf_tgt app. As a result, all IOs to all rbd bdevs will be handled on
the same CPU core and will be a performance bottleneck.

This patch allows all cores to serve IOs for one image by its single
ImageCtx.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>
Signed-off-by: Allen Zhu <allenzhu@nvidia.com>
Change-Id: If2c7e67254eaa1184bb0369e22e4f1938d86ca1f