bdev/rbd: Loadshare rbd bdevs between CPU cores
In current implementation all IOs for bdev rbd are served by single
"main" thread. It is set to the thread that created the first IO
channel for this rbd bdev. In many SPDK apps it will be the same
thread for all rbd bdevs, e.g. the 'nvmf_tgt_poll_group_0' for
nvmf_tgt app. As a result, all IOs to all rbd bdevs will be handled on
the same CPU core and will be a performance bottleneck.

This patch adds loadsharing of rbd bdevs between CPU cores. This is
achieved by creating bdev_rbd thread per CPU core and later assigning
"main" thread from this list in round-robin fashion.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>
Signed-off-by: Allen Zhu <allenzhu@nvidia.com>
Change-Id: Ib0378068a11e3a39ead105454c836fd243399043