lib/mlx5: Support signature and crypto in same MKey
The current UMR configure functions allow setting the crypto or
signature attributes but not both. However, the HW can execute the
crypto and CRC operations simultaneously, which is a more optimal way
than running them one by one.

This commit adds a new UMR configure function that allows combining the
crypto and signature in the same MKey.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>