accel/mlx5: Allocate signature context for crypto device
Allocate a context for signature offload operations when CRC support is
enabled. This commit is a preparation for CRC32C and COPY_CRC32C
support. The configuration option that allows enabling CRC support is
added in the subsequent commits.

Once an MKey is created with crypto_en, the MKey cannot be used without
the crypto attributes. That's why we need another MKey pool for
signature operations.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>