accel/mlx5: Merge ENCRYPT and CRC32C tasks
Optimize execution of the CRC32C task when it is posted next to the
ENCRYPT task in the same sequence. The feature is disabled by default,
and the subsequent commit adds a configuration option to enable it.

The implementation has a limitation. A CRC task fails if it is not
merged into a crypto task. The limitation does not break the NVMe TCP
initiator, which is the only place where the feature is supposed to be
used.

There are different ways to remove the limitation. Each way adds
overhead, and it is unclear which one is better from the performance
perspective. So, benchmarking is needed to choose the best way, and it
will be done later.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>