lib/mlx5: Add wrapper for WQE Control flags
This commit does not have any functional change. It only adds a wrapper
for the WQE Control flags defined in rdma-core. The wrapper is needed to
introduce a new flag in the subsequent commits. The new flag is missing
in rdma-core but can be added in the future. So, we add SPDK_ prefix to
avoid potential conflicts with future rdma-core versions.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>