accel/mlx5: Use different wrappers for sig and crypto mkeys
We need to extend the wrapper structure to handle signature error
completions. However, the upcoming extension is useless for crypto. So,
let's use different wrapper structures for crypto and signature MKeys to
avoid growing the memory footprint for crypto MKeys.

This commit is a preparation with no functional change and only renames
the wrapper structures.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>