accel/mlx5: Convert flags in task structure into bit field
Features added in the next commits require new flags in the task
structure. Since we want to keep the structure as small as possible,
adding new flags as bool variables is a bad idea. Let's convert the
existing flag into a bit field to simplify adding the new flags in the
subsequent commits.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>