accel: Add CRC32C check operations
Add a new opcode to check CRC32-C. The input parameters for the new
operation are an IO vector and an expected CRC32-C value. The operation
succeeds if the given CRC32-C is a valid CRC for data the IO vector
points.

HW limitations do not allow offloading the existing CRC operations in
some cases. The new opcode can be used to offload CRC calculation in
those cases.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>