accel/mlx5: Handle error state in PSV
Once a signature error occurs, the PSV object is moved to the error
state. SET_PSV WQE must be executed to reset the error before reusing
this PSV.

This commit extends the PSV structure to add an indication of the error.
It also implements handling of the error. Handling of SIGERR CQE that
leads to the error state in PSV will be added in subsequent commits.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>