lib/mlx5: Add raw_data_size to signature attributes
The current implementation assumes that BSF.raw_data_size is equal to
MKeyCtx.len. The assumption is only correct for the cases when the
signature exists on the wire domain only. Let's extend the signature
attributes to support cases with signatures on the memory domain. It is
needed to support signature check operation.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>