accel/mlx5: Simplify counting of work requests posted to device
One request on the task causes some number of posted WQEs to SQ. The
number depends on several factors, so the calculation looks complex. The
changes in the next commits can make the calculation even more complex.
To drop the complex calculation, let's add a counter to the task
structure.

There are two different entities called requests. The first one is a
part of the work in the task. The second one is a WQE posted to device.
It looks confusing. Let's rename the second one to Work Request to avoid
misunderstanding.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>