accel/mlx5: Reset merge flag properly
We must reset the merge flag when the main task is failed for any
reason, not only when a signature error is detected. Add the missing
check to accel_mlx5_task_fail() to fix that.

When there is a signature error in the crc32c + decrypt task, we must
reset the merge flag and accel_mlx5_crc_decrypt_task_complete() is the
right place to do that.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>