nvme_rdma: Flush recv WRs when qp is removed from outstanding
When a request completes, we queue recv WR and increment
current_num_recvs counter. This counter is used to check
if qpair is clean and can be destroyed - we expect to
reap completions for all submitted WRs once qpair is
in error state. But real flush of queued recv WRs to
the HW is handled by poll group and only happens when
qpair is the outstanding_qpairs. As result, we don't
reap recv completions and qpair wastes time in
LINGERING state.
Fix is:
1. Change the order of queuing recv WR/
completing request in nvme_rdma_request_ready
2. Cache cpl to avoid recv WR and cpl reuse while
SW processes completion
3. Flush recv WRs in nvme_rdma_req_complete if
qpair is removed from the outstanding_qpairs list

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: If806796e48549d9f84db01359de33d377f76d34c