bdev/nvme: Workaround for ns reconstruct when target restarted
In one of our systems, the following behavior was observed for NVMe-oF
TCP.

While FIO runs on the host, restart the target. Then, the initiator
succeeded to reconnect to the target. However, the initiator received
an empty ns list and deleted namespaces. Then, before completing reset,
the initiator received an async ns change event and created namespaces.
After reset completes, IOs resubmission started. However, IO path still
kept pointers to deleted namespaces. Hence, segmentation fault occurred.

We are still investigating why an empty ns list was received for
identification. However, segmentation fault is a critical error and
should be fixed.

Updating nvme_ns->ns can be a minimal workaround and is not harmful for
normal cases. Hence, add a workaround with a log to notify this case
occuured.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>