accel/mlx5: Allow only 1 crypto dev if --allowed-devs is empty
Crypto key is created for every device which supports crypto due
to mlx5 library is not aware of devices which are actually used.
If the used doesn't specify --allowed-devs, we select a device
which supports most of offloads. In that case crypto key still
uses all available devices. To fix that, allow only the best
dev if it supports crypto

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I786a7046a3cfd2410915428c04237f640912792f