lib/init: do not fail on missing RPC when stop_on_error is not set
SPDK apps support --json-ignore-init-errors option, to allow
continuing past RPCs that might fail.

But this was only working for RPCs that actually exist, and
then fail execution. It did not work for RPCs that didn't exist
at all. For example, if you save a JSON config from a running
spdk_tgt, it will include a lot of RPCs (nvmf, iscsi) that
don't exist in an app like bdevperf.

So fix up app_json_config_load_subsystem_config_entry() so that it
just continues to the next RPC if it finds an RPC call that doesn't
exist.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I60e6b55c7eda088725bf7486d3f67eeb5ddc3538
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22776
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot