lib/dma: Add API to get user context
When a memory domain is created, the user can
pass a pointer to a context which contains another
pointer to a domain-specific context (aka user
context). For RDMA memory domain this user context
contains a pointer to a protection domain.
Further patches will use this user context in IO
path, but with current API we'll have to dereference
two pointers whcih isn't a performance way.
To have less pointer dereferences, embed the user
context at the end of memory domain structure and
add API to get opaque pointer to the context

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I55c247e21c6e990dc3c2d7a02b2ff6e73548a898