sock: Remove zero copy APIs
Nothing currently uses these. We plan to introduce a different approach
for zero copy rx in the patches following this one.

As part of this we need to shift the buffer pool down into the io_uring
sock module. It is the only module that will need a pool like this.

Change-Id: Icbcc9a895a920851e20a92996e6ceded8bde804f
Signed-off-by: Ben Walker <ben@nvidia.com>