bdev: Swap bdev QoS and group QoS in qos_queue_io()
Swap quota operation of bdev QoS and group QoS in qos_queue_io() to
fix possible livelock. If group QoS is done first and then bdev QoS next,
other bdevs can be blocked.

Rewind operation will be removed in the following patches. We do not have
to keep related comments. Remove the comments in this patch for simplification.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>