nvme: Shift work in .poll_group_remove to .ctrlr_disconnect_qpair
For each of the transports, do any remaining work when the qpair is
disconnected rather than when the qpair is removed from the poll group.

We now see that all of the work of adding and removing a connection from
a group is best handled when the qpair is connected and disconnected.

Change-Id: I75492aa71e48be18e95cde666b6fe513c45891f2
Signed-off-by: Ben Walker <ben@nvidia.com>