bdev: Use node list to manage group and QoS instead of ref. count
We used ref. count to manage group. However, it was wrong. If a bdev is
opened before the bdev is added to a group and then the bdev is added
to the group, group is closed without open. Then ref. count goes to
negative.

Instead of group open/close, we can start group QoS when a bdev is newly
added to the group.

This simplifies the bdev group implementation a lot.

Additionally fix a potential bug in this patch. bdev_group_remove_bdev()
may be called on a different thread from bdev_group_add_bdev() is called.

The following patches will verify these fixes using unit tests.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>