lib/mlx5: Fix CQ resize
We create a CQ with IBV_CREATE_CQ_ATTR_IGNORE_OVERRUN to avoid
updating the consumer index in the poll function. It causes an
error during the resize operation when checking that the new CQE
size equals or exceeds the number of outstanding CQEs. The check
fails because the consumer index is never updated, and the number
of outstanding CQEs is incorrectly calculated. Let's update
the consumer index before resizing the CQ to fix the error.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>