nvmf: Link doca_sta statically
The doca_sta library contains a binary that is loaded to the DPA,
and the DPA does not allow loading binaries without a signature
when secure boot is enabled. So far, DOCA SDK does not support
the signature for dynamic libraries. So, using the static library
is the only way to run the NVMF RDMA Target Offload feature on
a DPA with an enabled secure boot.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>