nvmf/rdma_offload: Fix errors during deleting offloaded NVMe IO queue
The NVMe CQ is polled during the removal to handle all outstanding
completions. Completions produced by the offload have no context in
the software and cause errors.

Setting no_deletion_notification_needed is not enough to solve the
issue because we still reach assert(0) in
nvme_pcie_qpair_process_completions().

Clear the NVMe CQ before the removal to avoid errors.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>