fsdev: Add notifications API
Notifications is a mechanism to deliver a filesystem event from fsdev
to the user. It is similar to events mechanism but there are a few key
differences:
- notifications should be explicitly enabled
- notifications may carry data
- notifications may have a reply and reply may carry data

This is the main motivation to introduce new API for notifications
instead of reusing existing events callback.

Notifications are enabled with `spdk_fsdev_enable_notifications()`
API. Notifications can be enabled only once for fsdev. This means
there can be only one consumer of notifications. If there were more,
it is problematic to aggregate notification replies from multiple
consumers.

Type of notification strictly defines the data provided along with
it and data required in notification reply.

This commit introduces two notifcation types: invalidate data and
invalidate entry. Both may have a reply with notification handling
status.

Signed-off-by: Evgeniy Kochetov <evgeniik@nvidia.com>