lib/rdma_provider: Add config to enable offloads on qp
When support_offload_on_qp is set, all qpairs are created
twice bigger than requested to accomodate additional WQEs.
Memory domain is configured with a pointer to qp which
allows to pass it to other modules.
CQ size remains unchanged, by design all UMR WQEs are
unsignaled, qp is configured to not generate CQE for
error WQE except of the first one, so we can keep
CQ of the requested size.

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I4965592351855fff177498f4cf3b0c08c7ccc487