lib/mlx5: Fix byteorder for invalidate_rkey
The byte order of the immediate field of the send WQE depends on
the opcode. The field must be in the big-endian byte order when
the opcode is MLX5_OPCODE_SEND_INVAL and in the host byte order
when the opcode is MLX5_OPCODE_SEND_IMM. The argument of
mlx5_qp_send() was also renamed to make adding
spdk_mlx5_qp_send_imm() simpler if we decide to add it.

Signed-off-by: Sergey Gorenko <sergeygo@nvidia.com>