fsdev: per-IO optional cleanup callback
This is called prior to a request being freed. This allows modules to
cleanup any resources they may have allocated while processing this
request.

The moment the fsdev_io buffer is provided by an API user, we won't be
able to use it after the completion callback is called, as the fsdev_io
could be reused. This is the reason that the cleanup_cb_fn does not
receive fsdev_io as a parameter but rather has a void *cb_arg attached
to it.

Change-Id: I5e174445a67d1c8e4a30b50e6215f27e3ba27d16
Co-authored-by: Ben Walker <ben@nvidia.com>
Co-authored-by: Anton Nayshtut <anayshtut@nvidia.com>
Signed-off-by: Ben Walker <ben@nvidia.com>
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>