lib/log: spdk_vlog() fixed
The ap argument list cannot use be re-used (see below), so we now copy
it in advance to allow the secondary vasprintf invocation if necessary.

From man vsnprintf:

The functions vprintf(), vfprintf(), vsprintf(), vsnprintf() are
equivalent to the functions printf(), fprintf(), sprintf(), snprintf(),
respectively, except that they are called with a va_list instead of a
variable number of arguments. These functions do not call the va_end
macro. Because they invoke the va_arg macro, the value of ap is
*undefined* after the call.

Change-Id: Ibcb2d1318c402bd70817deedf76018a64d596752
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>