fsdevperf: add fsdevperf application
This app will allow users to test the performance of fsdevs.  It's the
equivalent of bdevperf, but for fsdev.

Initially, it's going to support only a single job specified on the
command-line, but the code is written in a way to easily support
multiple jobs.

For each job, users must specify a path to a file to submit IO to.  The
first component in that path is the name of the fsdev that the file
resides on.  For instance, /foo/bar describes a file called "bar" on an
fsdev called "foo".  Currently, this file must exist before starting up
the app, but later on, an option will be added, so that fsdevperf itself
will create it.

Also, for now, the app only opens up an fsdev, mounts it, opens the
file, closes it, and unmounts fsdev.  Submitting IO, along with other
features, will be added in following patches.

Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>
Change-Id: Ie3e1e638429e015cfeab6fb067699992bf1dd2e6