fsdevperf: submit I/O
The app will now submit I/O.  Currently, it's always sequential reads,
but users can specify the size of each I/O, the queue depth, and the
total number of bytes to be read.  If that total is less than the size
of the file, the app will start reading from the beginning.

Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>
Change-Id: I59aaaea42b39a275f07a31ba086b14f80fc58f4f