app: promote group_id to unsigned long long
SPDK_TRACE_MAX_GROUP_ID was increased to 32, so the shift can now
overflow an int, so promote group_id to unsigned long long to avoid
that.  Theoretically, it only needs to be promoted to unsigned long, as
the maximum value of group_id is 31, which is fine for an unsigned long,
but, since tpoint_group_mask is uint64_t, promote group_id to 64 bits
too.

This doesn't happen in upstream, because the increase was only done in
our fork.

Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>