lib/rmem: lock removed
The rmem is a low-level abstraction and has no knowledge of how it is
being used. As we see now, it's used either in a context that doesn't
require locking (init, recovery and the like) or as part of a bigger
critical section (like fobject/fhandle addition by the fsdev aio).

Thys, we removed the lock to avoid unnecessary / double locking.

It makes an API user responsible for thread-safe usage of the rmem API.

Change-Id: I13e136faf5978b078afb764dbe09fb1369271358
Signed-off-by: Anton Nayshtut <anayshtut@nvidia.com>