fsdevperf: add internal file setup/cleanup jobs
These jobs are spawned before and after the workloads to prepare and
clean up the files.  For now this only includes getting and releasing
file handles, but it'll be extended in the future to create and allocate
the files.

Both of these jobs spawn exactly one task for each file.  The setup job
spreads these tasks across all available cores.  Currently this is
unnecessary, but once files are created as part of this job, it'll speed
up that process.

Change-Id: I76f9c026b14247d44bcf2e93fcb11109b96a4d7e
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>