sock: Remove thread library dependency
The sock library mistakenly gained a dependency on the thread library
when interrupt support was added. Unfortunately, this is not allowed.
The lib/nvme code is intended to be independent of the rest of the SPDK
framework, and lib/nvme depends on lib/sock for the TCP transport, so it
must not depend on lib/thread.

However, this is easily resolved by making lib/sock just return the
interruptfd instead of actually doing the registration itself.

Change-Id: Idb75a035967582aa27f98b3db1bec6fa889a9833
Signed-off-by: Ben Walker <ben@nvidia.com>