fsdevperf: generate file names
If job's path is set to fsdev's root (e.g. /fsdev0), we'll now generate
file names (in the form of <job>.<core>) for each core to submit I/O to.
These files are treated the same way as other files (i.e. if they exist,
they're not recreated, they're not unlinked at the end, etc.).

Change-Id: I3cb905f2d8f309447704433223ffab15995b46a4
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>