fsdevperf: allocate single buffer for file setup
There's no need to allocate a separate buffer for each file, as we're
filling them all with the same data (zeroes) anyway.  And this can be an
actual problem when the number of files gets large.

This also allows us to skip buffer allocation for the cleanup job, which
is also completely unnecessary.

Change-Id: I247a945a5d04205556b0b6b68168884d19cd4c79
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>