rdma_provider: Add spdk_rdma_provider_memory_key_get_key()
In the nvmf/rdma transport, we need to find a mkey object whose
key ID matches, but only a opaque pointer to the mkey object is
available.

rdma_provider knows the definition of the mkey object. Hence,
rdma_provider provides a helper function to return the mkey of
the opaque mkey object.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>