fuse_dispatcher: move fuse_io->unique assignment up
This fixes completions for operations with opcodes that we don't
understand.  Previously, we'd complete them before the unique
assignment, which would result in sending some garbage value as the
unique id.

This does happen on newer (v6.6 and up) kernels, where the fuse driver
tries to send us FUSE_STATX commands, which we don't support.

Change-Id: I0fcb3ff027130518fa4bf92cc0b9da30b273b7cc
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>