fuse: process fuse requests
Requests are now received from the kernel and executed by
fuse_dispatcher.  Each request is associated with a 132k (128k for the
data plus one page for the headers) buffer.  This buffer is then chunked
into "in" headers and data, and "out" headers and data.

For most operations, we put fuse_in_header in iov[0] with the rest of
input in iov[1].  We do the same for the output: iov[0] contains
fuse_out_header, while the rest of output in iov[1].  We do this, as
opposed to using one iov for input and one iov for output, because some
of the fuse_dispatcher code has assumptions on the iov layout.  Thanks
to this scheme, we don't have to understand all FUSE operations. The two
exceptions are READ and WRITE operations.

For a READ, we also need to limit the size of the output buffer to
fuse_read_in.size.  That's because some fsdevs (e.g. aio) ignore the
size argument and will try to read as much as the size of the buffer
allows.

For a WRITE, fuse_dispatcher requires the data to be in a separate iov
from the headers, so we put both "in" headers (fuse_in_header and
fuse_write_in) in iov[0] and the data in iov[1].

Change-Id: Iba670a5cadd1a7534832cac5cc072e9d07554a31
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>