test: add helper script for running fstests
This script automates the steps necessary to run the (x)fstests [1]
suite with SPDK fsdevs via FUSE using examples/fuse/fuse app.  It can
also be used as a guideline for running these tests manually.  It allows
testing any type of fsdev, as the user specifies the configuration of
the fuse app and the name of the fsdev to use for testing.

It does two things: creates xfstests configuration and installs the
mount program, mount.fuse.spdk to /sbin.  The latter allows fstests
scripts to do:

  # mount -t fuse.spdk fsdev0 /mount/path

This mount program is a simple shell script that starts up the fuse app
to mount requested fsdev at a specified mount point and wait until the
mount is visible in the filesystem.

The script receives two types of arguments: its own parameters (e.g.
path to the xfstests repo, arguments to the fuse app) and parameters
passed to the xfstests check script, including the type of tests to
execute.  All arguments after "--" are passed to the xfstests check
script.

For instance,

  # test/fsdev/fstests.sh -x /path/to/xfstests-dev -c fsdev.json \
       -d fsdev0:/mnt/foo -- -fuse -g quick

will run tests from the "quick" group against fsdev "fsdev0", which will
be mounted at /mnt/foo.

[1] git://git.kernel.org/pub/scm/fs/xfs/xfstests-dev.git

Change-Id: I9f57acd92b4936840aa7ef12e98c1af4e64b62e1
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>