fuse, fuse_dispatcher: don't reply to FUSE_INTERRUPT requests
FUSE_INTERRUPT should either cause the original request to be aborted
and completed or, if that's not possible, it should be ignored.  In both
cases, a reply for the interrupt request is not needed.  In fact, if the
original request has already been completed, a reply to FUSE_INTERRUPT
isn't expected and will cause the driver to return an error.

There are two cases where the reply can be used: to notify the driver
that userspace doens't support interrupts (error=ENOSYS) or that it
wants to requeue the interrupt (error=EAGAIN).  Since we don't need
either of these functionalities, we always skip the reply.

Change-Id: Ie3057707c46f5fe433a901999d63ba26610578e4
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>