fuse: suppress common error messages
When an fsdev is unmounted outside of the process (e.g. via umount(1)
from shell), reads from the fuse device will fail with ENDOEV.
Similarly, umount2(2) will also fail in this case, as the filesystem is
already unmounted in this case.  Since this not an error, suppress the
error messages printed when this happens.

Change-Id: Idf9048ee674cc43ac8e8974e83005ab0d52e4884
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>