fsdevperf: fix scan-build warning
scan-build complains about the value never being used:

fsdevperf.c:1141:3: warning: Value stored to 'rc' is never read [deadcode.DeadStores]
 1141 |                 rc = -EINVAL;
      |                 ^    ~~~~~~~

It happens, because scan-build is built in debug mode, so the assertion
just after this assignment causes an abort() which never returns.
Moving the assertion before the assignment seems to silence this
warning.

Change-Id: I91eaa3fb791a1c3ae9842e1cabd398a165b6f2fd
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>