nvme_rdma: Skip appending copy operation for nvme_write_zeroes
We appended accel copy operartion for NVMe write zeroes command and
it caused unexpected SGL check error when processing the copy operation.

First of all, accel copy operation is not necessary to submit
NVMe write zeroes command to target.

As a bug fix, skip appending copy operation if payload size of the
request is zero. This can cover more opcodes rather than checking if
opcode is not NVMe write_zeroes.

Probably we can limit appending copy operation only to NVMe read or
write, but we can do it later.

Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>