bdev: Fix qos destroy race condition
When a QoS handle is destroyed and it was
the last QoS handle in the global list, a
QoS poller is stopped. But an upper layer
could call spdk_bdev_finish earlier, this
function destroyed QoS poller as result we
dereference a NULL pointer in spdk_poller_pause
Also, since _bdev_qos_destroy is called async-ly
via thread_send_msg, spdk_bdev_finish can be called
after sending a message but before processing it.

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I3e083097f72b66986454534a02054621ee0ab968