fsdev/aio: fhandles no longer hold pointers to fobjects
Instead, in the one place it is needed (during unref), we dynamically
lookup the fobject instead.

This is part of an important change to never hold pointers to fobjects
across yield points (pollers). We want to always look up the fobject,
use what we need, and discard the pointer immediately. Once all code
uses this pattern, we can reduce the amount of reference counting
necessary and improve the performance of some code paths such as unlink.

Change-Id: I1cce7d6d8a306668e6c626786ad7a7dac26d0ec8
Signed-off-by: Ben Walker <ben@nvidia.com>