ci: added coverity stage
Added a new stage for Coverity analysis in the CI pipeline.
This stage includes the necessary scripts and configurations to run Coverity scans,
ensuring code quality and security checks are integrated into the development workflow.
The changes also include updates to the job matrix to accommodate the new Coverity stage.

Signed-off-by: Iaroslav Sydoruk <isydoruk@nvidia.com>