fuse_dispatcher: add support for FUSE_DIRECT_IO_ALLOW_MMAP
This is required to support mmap on files in a fs mount.

The kernel allows us to say we support this flag, even though it was added in
fuse v7.39, and we are only reporting v7.34.

Note that fuse_init now has a 'flags2', which is where FUSE_DIRECT_IO_ALLOW_MMAP
lives, so we need to also add checks for FUSE_INIT_EXT to see if the kernel
even passed us anything in flags2.

Signed-off-by: Jim Harris <jim.harris@nvidia.com>